refactor: combine story-related tests in the groups
