+++ test/app/features/stories/widgets/story_progress_tests.dart
@@ -5,8 +5,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:ion/app/features/feed/stories/data/models/story.c.dart';
 
 import '../../../../fixtures/factories/post_factory.dart';
+import '../../../../robots/stories/image_progress_robot.dart';
 import '../../../../robots/stories/story_progress_bar_robot.dart';
 import '../helpers/story_test_models.dart';
+import '../helpers/story_test_utils.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -14,7 +16,6 @@ void main() {
   registerStoriesFallbacks();
 
   const viewerPubkey = 'alice';
-
   final viewerStories = UserStories(
     pubkey: viewerPubkey,
     stories: [
@@ -30,7 +31,6 @@ void main() {
         stories: [viewerStories],
         viewerPubkey: viewerPubkey,
       );
-
       robot.expectSegmentCount(2);
     });
 
@@ -40,10 +40,8 @@ void main() {
         stories: [viewerStories],
         viewerPubkey: viewerPubkey,
       );
-
       robot.advance();
       await tester.pump();
-
       robot
         ..expectSegmentState(index: 0, isCurrent: false, isPrevious: true)
         ..expectSegmentState(index: 1, isCurrent: true, isPrevious: false);
@@ -63,18 +61,49 @@ void main() {
         stories: [viewerStories, bobStories],
         viewerPubkey: viewerPubkey,
       );
-
-      // advance to s2 (same user) and then to bob/b1
       robot
         ..advance()
         ..advance();
       await tester.pump();
-
       robot
         ..expectViewerState(userIndex: 1, storyIndex: 0)
         ..expectSegmentCount(2)
         ..expectSegmentState(index: 0, isCurrent: true, isPrevious: false)
         ..expectSegmentState(index: 1, isCurrent: false, isPrevious: false);
     });
+  });
+
+  group('Image progress & keyboard (Robot)', () {
+    testWidgets(
+      'progress pauses on keyboard open and resumes after close',
+      (tester) async {
+        final post = buildPost('img_keyboard');
+        final robot = ImageProgressRobot(tester, post: post);
+
+        await pumpWithOverrides(
+          tester,
+          child: robot.buildHost(),
+          overrides: storyViewerOverrides(post),
+        );
+        await robot.attach();
+
+        robot.markImageLoaded(); // preload done
+        await robot.wait(const Duration(seconds: 2));
+
+        // Act 1 – open keyboard (= pause)
+        robot.paused = true;
+        await robot.wait(Duration.zero);
+        final frozen = robot.progress;
+
+        await robot.wait(const Duration(seconds: 1));
+
+        robot
+          ..expectFrozen(frozen)
+          ..paused = false;
+        await tester.pumpAndSettle(const Duration(seconds: 4));
+
+        robot.expectCompleted();
+      },
+    );
   });
 }
